Make raceIds `uint64`

+++ contracts/interfaces/IExotic.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 interface IExotic{
 
     struct Bet {
-        uint256 raceId;
+        uint64 raceId;
         uint256 amount;
         address account;
         uint8 prediction;
@@ -21,9 +21,9 @@ interface IExotic{
 
     function endRace(uint256 requestId, uint256[] memory randomWords) external;
     function userBetCount(address userId) external view returns (uint256);
-    function race(uint256 raceId) external view returns (Race memory);
+    function race(uint64 raceId) external view returns (Race memory);
     function frequency() external view returns (uint256);
     function userBet(address user, uint256 betId) external view returns (Bet memory);
-    function raceResult(uint256 raceId) external view returns (uint256[1] memory);
-    function totalWagered(uint256 raceId) external view returns (uint256);
+    function raceResult(uint64 raceId) external view returns (uint256[1] memory);
+    function totalWagered(uint64 raceId) external view returns (uint256);
 }
